+++ src/test/java/study/step1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step1;
 
 import static org.assertj.core.api.Assertions.*;
