+++ src/test/java/study/step2/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,6 +8,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import study.step2.StringAddCalculator;
 
 public class StringAddCalculatorTest {
